DRIVERS-1652: Clarify events.json and result.json produced by workload executor (#120)
* DRIVERS-1652: Clarify JSON produced by workload executor

This rewrites portions of the workload executor spec to clearly define how events.json and results.json are produced. This also improves instructions for when the test runner propagates and error/failures, which must then be reported by the workload executor in the same format.

* DRIVERS-1652: Make ValidateWorkloadExecutor consistent with spec

Removes the placeholder stats if results.json cannot be read. Per the spec, the workload executor should always produce results.json (and events.json).

Always assert presence and types of required fields in results.json and events.json. Improve events.json assertions for errors and failures.

Add numErrors-as-failures test.

* Always remove events.json before spawning workload executor

* Fix wrong comment wording

Co-authored-by: Andreas Braun <alcaeus@users.noreply.github.com>